Ensure abilityScoreToModifier returns an int

+++ src/DndBeyond/Rules.php
@@ -6,6 +6,6 @@ class Rules
 {
     public static function abilityScoreToModifier(int $abilityScore): int
     {
-        return floor(($abilityScore - 10) / 2);
+        return (int) floor(($abilityScore - 10) / 2);
     }
 }
